Create cfl-website-article-grid
A content block used in Braze for the Canadian Football League that render's a 2 column grid of articles with a specific tag or category, including exclusions from cfl.ca if more than 1 post is returned and renders 1 column if 1 post is returned. This does nothing if no posts are returned.